<shake> command is added, based on move
I need this to implement renpy's "with hpunch" / "with vpunch". Default
values of period and duration are taken from renpy sources. At the very
begining of shaking image is moved by dx, but it's exactly renpy's
behaviour, so i decided not to change it to smooth starting.

You can test it with something like this:

<shake asset="backgrounds" dx="-15px" />
<shake asset="backgrounds" dy="5%" period="80"/>
<wait />